moved buildFunctions inside build.tsx
